refactor: use memmap in hello world
