+++ jina/helloworld/fashion/my_executors.py
